+++ types/index.d.ts
@@ -1,35 +1,47 @@
+import express = require('express');
+
 declare module '@uttori/plugin-upload-multer';
 
-declare module "index" {
-    export = MulterUpload;
-    class MulterUpload {
-        static get configKey(): string;
-        static defaultConfig(): object;
-        static validateConfig(config: {
-            configKey: object;
-        }, _context: object): void;
-        static register(context: {
-            hooks: {
-                on: Function;
-            };
-            config: {
-                events: object;
-            };
-        }): void;
-        static bindRoutes(server: {
-            post: Function;
-            use: Function;
-        }, context: {
-            config: {
-                directory: string;
-                route: string;
-            };
-        }): void;
-        static upload(context: {
-            config: {
-                directory: string;
-                route: string;
-            };
-        }): Function;
-    }
+export class MulterUpload {
+  static get configKey(): string;
+
+  static defaultConfig(): MulterUploadConfig;
+
+  static validateConfig(config: {
+        configKey: MulterUploadConfig;
+    }, _context: object): void;
+
+  static register(context: {
+        hooks: {
+            on: Function;
+        };
+        config: {
+            events: object;
+        };
+    }): void;
+
+  static bindRoutes(server: {
+        post: Function;
+        use: Function;
+    }, context: {
+        config: {
+            directory: string;
+            route: string;
+        };
+    }): void;
+
+  static upload(context: {
+        config: {
+            directory: string;
+            route: string;
+        };
+    }): express.RequestHandler;
 }
+
+export type MulterUploadConfig = {
+    events?: object;
+    directory: string;
+    route: string;
+    publicRoute: string;
+    middleware: express.RequestHandler[];
+};